smartscales/requirements.txt
absl-py==0.7.1
astor==0.8.0
awscli==1.16.193
Babel==2.7.0
boto3==1.9.183
botocore==1.12.183
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
docker-pycreds==0.4.0
docutils==0.14
dominate==2.3.5
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Moment==0.8.0
Flask-SocketIO==4.1.0
Flask-WTF==0.14.2
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==8.7
keyrings.alt==3.1.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
networkx==2.3
numpy==1.16.4
opencv-python==3.4.2.16
Pillow==5.3.0
protobuf==3.8.0
pyasn1==0.4.5
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.8.2.post1
python-socketio==4.2.0
pytz==2019.1
PyWavelets==1.0.3
pywin32-ctypes==0.2.0
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
ruamel.yaml==0.15.97
s3transfer==0.2.1
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tabulate==0.8.3
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
urllib3==1.25.3
visitor==0.1.3
voluptuous==0.11.5
websocket-client==0.56.0
Werkzeug==0.15.4
WTForms==2.2.1
